Requirements fixed, Controller JSON fixed
Controller.change_traffic_light now works with JSONs.

End nodes do not have traffic light and causes problem. Temporarily
fixed by returning so that if a nodes number of intersection == 1, ie it
is an end node, nothing happens.

NOTE: Future work requires nodes to either initialise with a traffic
light that can be changed to red or green, or all end nodes have a green
traffic light which cannot be changed. Without this, we are wasting
state space without any benefit.